src/courses/django/chat/requirements.txt
Django==4.0.6
pytest==7.1.2
celery==4.4.7
djangorestframework